Extract ColorSwabs.
