+++ src/client/components/color-swabs/color-swabs.scss
@@ -14,22 +14,40 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 @import '../../imports';
 
-.hover-multi-bubble {
-  @extend %menu-cont;
-  background: rgba($white, 0.9);
-  position: absolute;
-  padding: 8px;
+.color-swabs {
+  table-layout: fixed;
 
-  .bucket {
-    color: $text-lighterish;
-    margin-bottom: 6px;
-    white-space: nowrap;
+  .color-swab {
+    width: 15px;
+    height: 15px;
+    border-radius: 2px;
   }
 
-  .segment-action-buttons {
-    margin-top: 7px;
+  .color-name {
+    @include ellipsis;
+    max-width: 140px;
+    padding-left: 5px;
+    line-height: 18px;
+  }
+
+  .color-previous,
+  .color-value {
+    line-height: 18px;
+    white-space: nowrap;
+    text-align: right;
+  }
+
+  .color-previous {
+    padding-left: 5px;
+    color: $text-lighterish;
+  }
+
+  .color-delta {
+    line-height: 18px;
+    white-space: nowrap;
+    padding-left: 5px;
   }
 }
+
